Rename typing to scalabel_typing to avoid collision with standard typing library
