+++ scalabel/label/__init__.py
@@ -1,3 +1,3 @@
 """Label definition and conversion."""
 
-from . import coco_typing, from_coco, io, to_coco, typing
+from . import coco_typing, from_coco, io, scalabel_typing, to_coco
